IBIS: Avoid assigning an uninitialised variable
This would be OK if m_reporter were overwritten in the
derived ctor after being filled with junk in the base
ctor. But this is not always true - the null ctor doesn't
do it. So it can stick about and fool the null check.

Rather than just doing that, push the reponsibility for
maintaining the m_reporter invariant contract (if it isn't null,
you may use it) to the base class ctor.

And get rid of the shadowing m_reporter in KIBIS.