Reference image: add a transform origin
This is stored as an offset from the image position,
which is the centre of the image.  When interactively edited, the
transform origin stays in place. This makes it easier to scale an image
around some fixed feature in the image.

For now this is not saved out to the format, so it
reinitialises to (0, 0) each time. This means scale around
the midpoint, which is the current behaviour.