NETLIST_PARSER: Put in own header
This one isn't as invasive as the others, but it is still
an implementation class that's not needed for public use
of netlists.

Actually, this could even be local to kicad_netlist_reader.cpp
as it's literally only used there.

Now, dsnlexer.h is exposed to only the files it needs, rather
than nearly 800 files.