Fix uninitialised vars in ZONE
ZONE has a huge number of state variables. Some of these
have setters that check m_foo == aNewFoo. This is only
good karma if m_foo is initialised!

This doesn't set all of the members (maybe it should,
or maybe there should be sub-objects with explicit ctors)
but at least it should prevent obvious un-inited reports.