Eeschema, fix possible crash when reading a .kicad_sch file (dangling pointer)
SCH_IO_KICAD_SEXPR_PARSER::parseSheet() can create a dangling pointer when reading
a property if the font used by the property text is not the default font: the
initial SCH_FIELD item is a temporary item and leave a dangling pointer in
m_fontTextMap liast when deleted. This pointer is now removed from list.