Fabmaster: unify text settings, adjust rotations.
This is a little tricky, but this seems to be the simplest
way to approximate the right text placement.

There are still some minor offsets and text size adjustments,
which might be possible to compensate for by experimentation.

(cherry picked from commit fee1d115a6479e89f39e6afefa86d1ef0cc09ebd)