Show save modified dialog on schematic sheet file rename.
Renaming a schematic sheet file can not be undone so it doesn't go
through the commit system.  This meant that the schematic modified
flag was not getting set.  If the only edit performed was renaming
a sheet file, the save modified flag would not get displayed when
closing the schematic editor.