Pcbnew: dimension suppress zeros by default
Per ASME-Y14.5-2009 section 2.3, trailing zeroes are not used for metric
dimensions, which are probably most common. Unit suffixes are also not
used in every dimension when specified as a global default, which is
also common (usally in the fab notes).

This appears to be common defaults in other EDA software.  This is also
compatible with ISO 129-1 section 4.3.

In Pcbnew, this is only the default until the user sets it in the board
settings, so existing designs will always keep their set values.