cli: add `--drill-shape-opt` to `pcb export dxf`
ADDED: add support for specifying drill shape ("drill marks" in the UI)
on the CLI when exporting DXF files of PCBs.

This standarizes on "actual size" for the drill marks, as do PDF and SVG
exporters, while the current, undefined and undocumented default,
appears to be "small marks".

Oddly enough, this option appears to only have effect in `--mode-new`.
I would appreciate guidance on what to do about this fact.