Revert "Pcbnew: reduce snap timeout a fraction"
Now the updating is fixed in d90862b44454a146ede30fe061002f6c0e8ceca9,
the hover timeout is shorter because it's not mistakenly waiting a
random-ish time for the next mouse movement. So this adjustment is not
required. In fact, if anything, it may be better to go the other way.

This reverts commit 7080d99464eb8ee72863ef24f1c71523dbbebb4b.