Ensure that the group RTree positions are updated
Group selections are typically only the group (and not the children), so
calling an RTree update on the selection will not update the positon of
the children, only the group.  This will lead to misses when calculating
hits for various operations until the elements are updated (usually in a
commit).  This is, however, too late for checks that happen during moves

Fixes https://gitlab.com/kicad/code/kicad/-/issues/20761