Fix update symbol from library when alternate pins are in play.
ADDED: An option in the update symbol dialog to reset alternate pins to
       their default value.

Don't set the symbol alternate pin when the updated symbol no longer
contains the alternate pin name.

Clear the alternate pin name when it's set to the default pin name.  This
is caused by a bug somewhere which hasn't been resolved and it causes the
library symbol comparison to fail when it's set.

This also changes the change/update symbol dialog layout.  Adding another
option check box was making the dialog too tall.  The option check boxes
were split into two columns which made the dialog wider.  It might be
worth using another wxCheckListBox for the options as more are added.