ADDED: Remote symbol placement
Allows activating a display panel that shows a webpage designed to
provide instances of symbols, footprints, 3d models and spice models to
the designer.

The webpage can be installed via PCM or directly in the panel.  Webpages
should be registered with the KiCad project in order to avoid a warning
when showing content.