Smarten libeval missing units warning.
In particular, don't choke on multiplicative statements
including a scalar.

The new algo is not actually smart enough to eval the
whole parse tree and separate conditionals from multiplication
from addition; it just contents itself with reporting
the common case of a single value with no units.

See comments in code.