DIALOG_LIB_SYMBOL_PROPERTIES: for fields Y pos, remove the sign handling.
This was a legacy thing. some time ago the Y axis of symbol graphics was
bottom to top, opposite of schematic Y axis.
This is no longer the case (since some time) and the Y sign change is a bug.

Note: for symbol files, the Y axis of symbol graphics is still bottom to top.