FONT_CHOICE: fix an issue specific to MSW
FONT_CHOICE use Clear() to delete the wxOwnerDrawnComboBox items on MSW.
however (wxWidgets bug?) it creates 2 issues: a wxWidgets alert and fires
OnTextCtrl events, creating annoying recursions because OnTextCtrl event
call Clear().
Looks like it did not happen on Linux.
So wxOwnerDrawnComboBox::Clear() is not used on MSW, and used on other platforms.